% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectortranslate.R
\name{get_default_osmconf_ini}
\alias{get_default_osmconf_ini}
\title{Get default osmconf.ini}
\usage{
get_default_osmconf_ini()
}
\value{
Path to the file
}
\description{
Returns the path to the \code{CONFIG} file used by this package when running the
\code{.osm.pbf} -> \code{.gpkg} conversion
}
\examples{
get_default_osmconf_ini()
}
