% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_geofabric.R
\name{gf_filename}
\alias{gf_filename}
\title{Get the filename of a file downloaded from geofabrik}
\usage{
gf_filename(name = "West Yorkshire",
  download_directory = gf_download_directory())
}
\arguments{
\item{name}{The name of the geofrabic zone. Must be an element in `geofabric::geofabric_zones$name`.}

\item{download_directory}{Where to download the data? `tempdir()` by default.
If you want to download your data into a persistent directory, set
`GF_DOWNLOAD_DIRECTORY=/path/for/osm/data` in your `.Renviron` file, e.g. with
`usethis::edit_r_environ()`.}
}
\description{
Get the filename of a file downloaded from geofabrik
}
\examples{
f = gf_filename("Isle of Wight")
f
file.exists(f)
}
