% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gf_find.R
\name{gf_find}
\alias{gf_find}
\title{Find geofabric zones}
\usage{
gf_find(name, ask = FALSE, max_dist = 5)
}
\arguments{
\item{name}{Name or spatial object of the geofabric zone to download}

\item{ask}{Should the user be asked before downloading the file?}

\item{max_dist}{What is the maximum distance in fuzzy matching to tolerate before asking
the user to select which zone to download?}
}
\value{
A data frame representing the matching items from the Geofabrik website
}
\description{
Find geofabric zones
}
\examples{
gf_find("Andorra")
}
