% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gf_find.R
\name{gf_find_sf}
\alias{gf_find_sf}
\title{Find geofabrik zones based on sf or sfc object}
\usage{
gf_find_sf(name, ask = FALSE, op = sf::st_contains)
}
\arguments{
\item{name}{String or \code{\link[sf:sf]{sf::sf}} spatial object of the geofabrik zone
to download. See examples.}

\item{ask}{Should the user be asked before downloading the file?}

\item{op}{The binary spatial predicate used to identify the smallest geofabrik zones
that matches the simple feature input in \code{name}}
}
\value{
A data frame representing the matching items from the Geofabrik website
}
\description{
Find geofabrik zones based on sf or sfc object
}
\examples{
name = sf::st_sfc(sf::st_point(c(0, 53)), crs = 4326)
gf_find_sf(name)
}
