% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_pbf.R
\name{make_additional_attributes}
\alias{make_additional_attributes}
\alias{make_ini_attributes}
\title{Get modified version of config file for reading .pbf files with GDAL/sf}
\usage{
make_additional_attributes(layer)

make_ini_attributes(attributes, layer,
  defaults = get_ini_layer_defaults(layer), append = TRUE)
}
\arguments{
\item{layer}{Which layer to read in? One of
"points"           "lines"            "multipolygons"    "multilinestrings" or "other_relations"}

\item{attributes}{Vector of character strings naming attributes to import}

\item{defaults}{The default attributes}

\item{append}{Should the columns named in `attributes` be appended to the default columns?
`TRUE` by default.}
}
\value{
A character vector representing an ini file.
}
\description{
This function provides a user friendly interface to the [GDAL OSM driver](https://gdal.org/drivers/vector/osm.html)
via `sf`. See https://github.com/r-spatial/sf/issues/1157 for details.
}
\details{
Default additional attributes were taken from https://taginfo.openstreetmap.org/keys
}
\examples{
\donttest{
make_ini_attributes("oneway", "lines")
}
}
