% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{oe_clean}
\alias{oe_clean}
\title{Clean download directory}
\usage{
oe_clean(download_directory = oe_download_directory(), force = FALSE)
}
\arguments{
\item{download_directory}{The directory where the \code{.osm.pbf} and \code{.gpkg}
files are saved. Default value is \code{oe_download_directory()}.}

\item{force}{Internal option. It can be used to skip the checks run at the
beginning of the function and force the removal of all \code{pbf}/\code{gpkg} files.}
}
\value{
The same as \code{unlink()}.
}
\description{
This functions is a wrapper around \code{unlink()} that can be used to delete all
\code{.osm.pbf} and \code{.gpkg} files in a given directory.
}
\examples{
# Warning: the following removes all files in oe_download_directory()
\dontrun{
oe_clean()}
}
