% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{oe_download_directory}
\alias{oe_download_directory}
\title{Returns the download directory used by the package}
\usage{
oe_download_directory()
}
\value{
A character vector representing the path for the download directory
used by the package.
}
\description{
This function returns the download directory used the package. By default, it
is equal to the output of \code{tools::R_user_dir("osmextract", "data")}. Such
value can be overwritten by setting the \code{OSMEXT_DOWNLOAD_DIRECTORY}
environment variable. If the directory does not exist, it will be created.
}
\examples{
oe_download_directory()

if (requireNamespace("withr", quietly = TRUE)) {
  withr::with_envvar(
    c("OSMEXT_DOWNLOAD_DIRECTORY" = tempdir()),
    oe_download_directory()
  )
}
}
