% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-network-graph.R
\name{oe_get_tidynetwork}
\alias{oe_get_tidynetwork}
\title{Obtain a tidy sf from OpenStreetMap data}
\usage{
oe_get_tidynetwork(
  place,
  mode = c("cycling", "driving", "walking"),
  ...,
  simplify_highway = TRUE,
  highway_filter = NULL
)
}
\arguments{
\item{place}{Description of the geographical area that should be matched with
a \code{.osm.pbf} file. Can be either a length-1 character vector, an
\code{sf}/\code{sfc}/\code{bbox} object with any CRS, or a numeric vector of coordinates
with length 2. In the last case, it is assumed that the EPSG code is 4326
specified as c(LON, LAT), while you can use any CRS with \code{sf}/\code{sfc}/\code{bbox}
objects. See Details and Examples in \code{\link[=oe_match]{oe_match()}}.}

\item{mode}{A character string of length one denoting the desired mode of
transport. Can be abbreviated. Currently \code{cycling} (the default), \code{driving}
and \code{walking} are supported.}

\item{...}{Additional arguments passed to \code{\link[=oe_get]{oe_get()}} such as \code{boundary} or
\code{force_download}.}

\item{simplify_highway}{logical, whether to simplify the highway values by removing the "_link" suffix and filtering by \code{highway_filter}}

\item{highway_filter}{character vector of highway types to keep, if \code{simplify_highway} is TRUE.
If not provided, all highway types returned by \code{oe_get_network} are kept. Valid values
for \code{highway_filter} are: "busway", "cycleway", "footway", "living_street", "motorway",
"path", "pedestrian", "primary", "residential", "rest_area", "service", "services",
"steps", "tertiary", "track", "trunk" and "unclassified".}
}
\value{
a \code{sf} object with standardised highway and oneway values
}
\description{
Obtain a tidy sf from OpenStreetMap data
}
\examples{

highway_filter = c(
 "motorway",
 "trunk",
 "primary",
 "secondary",
 "tertiary",
 "unclassified",
 "residential"
)

tidynet_sf <- oe_get_tidynetwork(place = "ITS Leeds", mode = "driving")
print(tidynet_sf)

tidynet_sf_filtered <- oe_get_tidynetwork(
  place = "ITS Leeds",
  mode = "driving",
  highway_filter = highway_filter
)
print(tidynet_sf_filtered)

}
\seealso{
\code{\link[=oe_get]{oe_get()}}, \code{\link[=oe_get_network]{oe_get_network()}}, \code{\link[=oe_get_sfnetwork]{oe_get_sfnetwork()}}, \code{\link[=oe_get_dodgrnetwork]{oe_get_dodgrnetwork()}}
}
