% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-transport.R
\name{oe_get_transport}
\alias{oe_get_transport}
\title{Import road network according to a specific mode of transport}
\usage{
oe_get_transport(place, mode = c("cycling", "walking"), ...)
}
\arguments{
\item{place}{Description of the geographical area that should be matched with
a \code{.osm.pbf} file. Can be either a length-1 character vector, an
\code{sf}/\code{sfc}/\code{bbox} object, or a numeric vector of coordinates with length 2.
In the last case, it is assumed that the EPSG code is 4326 specified as
c(LON, LAT), while you can use any CRS with \code{sf}/\code{sfc}/\code{bbox} objects. See
Details and Examples in \code{\link[=oe_match]{oe_match()}}.}

\item{mode}{The mode of transport. Can be abbreviated.}

\item{...}{IGNORED FOR THE MOMENT. Additional arguments passed to \code{oe_get()}.
Please note that the arguments \code{layer} and \code{vectortranslate_options} cannot
be provided since are set according to the \code{mode} parameter.}
}
\value{
An \code{sf} object.
}
\description{
TODO
}
\details{
TODO (precisely describe the functions)
}
\examples{
1 + 1
}
\seealso{
\code{\link[=oe_get]{oe_get()}}
}
