% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{oe_search}
\alias{oe_search}
\title{Returns a single sf feature associated with a search term}
\usage{
oe_search(
  query = "Leeds",
  base_url = "https://nominatim.openstreetmap.org",
  destfile = tempfile(fileext = ".geojson")
)
}
\description{
Documentation: wip, not yet exported
}
