% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectortranslate.R
\name{oe_vectortranslate}
\alias{oe_vectortranslate}
\title{Translate an OSM file with \code{.osm.pbf} format into \code{.gpkg} format}
\usage{
oe_vectortranslate(
  file_path,
  vectortranslate_options = NULL,
  layer = NULL,
  osmconf_ini = NULL,
  extra_attributes = NULL,
  force_vectortranslate = FALSE,
  quiet = TRUE
)
}
\arguments{
\item{file_path}{Character string representing the full path the .osm.pbf file}

\item{vectortranslate_options}{Options to pass to the \code{\link[sf:gdal_utils]{sf::gdal_utils()}}
argument \code{options}. Set by default. Check details and examples.}

\item{layer}{Which \code{layer} should be read in? Typically \code{points}, \code{lines}
(the default), \code{multilinestrings}, \code{multipolygons} or \code{other_relations}.}

\item{osmconf_ini}{The configuration file specifying which columns should be
in the resulting data frame. See documentation at
\href{https://gdal.org/drivers/vector/osm.html}{gdal.org} for details.}

\item{extra_attributes}{Which addition columns, corresponding to OSM keys,
should be in the resulting dataset? \code{NULL} by default.}

\item{force_vectortranslate}{Force the original \code{.pbf} file to be translated
into a \code{.gpkg} file, even if a \code{.gpkg} associated with the \code{provider} zone
already exists.}

\item{quiet}{Boolean. If \code{FALSE} the function prints informative messages.}
}
\value{
Character string representing the path of the .gpkg file.
}
\description{
This function is used to convert a \code{.osm.pbf} file into \code{.gpkg} format. The
conversion is performed using
\href{https://gdal.org/programs/ogr2ogr.html#ogr2ogr}{ogr2ogr} through
\code{sf::gdal_utils()}.
}
\details{
The .gpkg file is created in the same directory as the input .osm.pbf
file. The conversion is skipped if the function detects a file . This
behaviour can be overwritten with the parameter \code{force_vectortranslate}.
}
\examples{
# First we need to match an input zone with a .osm.pbf file
its_match = oe_match("ITS Leeds", provider = "test")
# The we can download the .osm.pbf files
its_pbf = oe_download(
  file_url = its_match$url,
  file_size = its_match$file_size,
  download_directory = tempdir(),
  provider = "test"
)
list.files(tempdir(), pattern = "pbf|gpkg")
its_gpkg = oe_vectortranslate(
 its_pbf
)
list.files(tempdir(), pattern = "pbf|gpkg")
}
