% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{oe_vtlist}
\alias{oe_vtlist}
\title{Convert a list to character strings for gdal_utils options}
\usage{
oe_vtlist(vtlist = NULL)
}
\arguments{
\item{vtlist}{}
}
\value{

}
\description{
Convert a list to character strings for gdal_utils options
}
\examples{
oe_vtlist(vtlist = list(overwrite = TRUE))
oe_vtlist(vtlist = list(overwrite = TRUE, f = "GPKG"))
oe_vtlist(vtlist = list(f = "GPKG", t_srs = "EPSG:32633"))
its = oe_get("ITS Leeds", quiet = FALSE, download_directory = tempdir())
# mapview::mapview(its[1:5, ]) # find suitable subset
its_sample_union = sf::st_union(its[1:5, ])
its_polygon = sf::st_convex_hull(its_sample_union)
its_poly_txt = sf::st_as_text(its_polygon)
q = oe_vtlist(list(clipdst = its_poly_txt))
its2 = oe_get("ITS Leeds", quiet = FALSE, download_directory = tempdir(),
  vectortranslate_options = q)
plot(its$geometry, col = "grey")
plot(its_polygon, add = TRUE)
}
