% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match.R
\name{osmext_check_pattern}
\alias{osmext_check_pattern}
\title{Check for patterns in the provider's data columns}
\usage{
osmext_check_pattern(
  pattern,
  provider = "geofabrik",
  match_by = "name",
  full_row = FALSE
)
}
\arguments{
\item{pattern}{Character string for the pattern that should be matched}

\item{provider}{TODO}

\item{match_by}{TODO}

\item{full_row}{TODO}
}
\value{
A
}
\description{
This function is used to explore the provider's data and check for patterns
in the existing columns
}
\examples{
osmext_check_pattern(
pattern = "Yorkshire",
provider = "geofabrik",
match_by = "name"
)
}
