% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{osmext_download}
\alias{osmext_download}
\title{Title}
\usage{
osmext_download(
  file_url,
  file_basename = basename(file_url),
  provider = "geofabrik",
  download_directory = osmext_download_directory(),
  file_size = NA,
  force_download = FALSE,
  max_file_size = 5e+08,
  verbose = FALSE
)
}
\arguments{
\item{file_url}{A}

\item{file_basename}{B}

\item{provider}{C}

\item{download_directory}{D}

\item{file_size}{E}

\item{force_download}{F}

\item{max_file_size}{G}

\item{verbose}{H}
}
\value{
ch
}
\description{
Title
}
\examples{
1 + 1
}
