% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{osmext_get}
\alias{osmext_get}
\title{Download, translate and read OSM extracts}
\usage{
osmext_get(
  place,
  layer = "lines",
  provider = "geofabrik",
  ...,
  match_by = "name",
  max_string_dist = 1,
  interactive_ask = FALSE,
  download_directory = osmext_download_directory(),
  force_download = FALSE,
  max_file_size = 5e+08,
  vectortranslate_options = NULL,
  osmconf_ini = NULL,
  extra_attributes = NULL,
  force_vectortranslate = NULL,
  osmext_verbose = FALSE
)
}
\arguments{
\item{place}{Description of the geographical area that should be download
through the chosen \code{provider}. Can be either a length-1 character vector, a
length-1 \code{sfc_POINT} object or a numeric vector with length 2. See details
and examples.}

\item{layer}{TODO}

\item{provider}{Which provider should be used to download the data? For the
moment we support only \href{http://download.geofabrik.de/}{\code{"geofabrik"}}.}

\item{...}{Arguments that should  be passed to \code{\link[sf:st_read]{sf::st_read()}}}

\item{match_by}{Which column of the provider data should be used for matching
the input place with the provider's data? The default is "name". Check
details and examples to understand how this parameter works. Ignored if
\code{place} is not a character vector since the matching is performed through a
spatial operation.}

\item{max_string_dist}{Numeric value greater or equal than 0. What is the
maximum distance in fuzzy matching to tolerate before asking the user to
select which zone to download? This parameter is set equal to 0 if
\code{match_by} is equal to \code{iso3166_1_alpha2} or \code{iso3166_2}. Check details and
examples to understand why this parameter is important. Ignored if \code{place}
is not a character vector since the matching is performed through a spatial
operation.}

\item{interactive_ask}{Boolean. If \code{TRUE} the function creates and
interactive menu in case the best match is further than \code{max_string_dist}.
Check details and examples to understand why this parameter is important.
Ignored if \code{place} is not a character vector since the matching is
performed through a spatial operation.}

\item{download_directory}{TODO}

\item{force_download}{TODO}

\item{max_file_size}{TODO}

\item{vectortranslate_options}{TODO}

\item{osmconf_ini}{TODO}

\item{extra_attributes}{TODO}

\item{force_vectortranslate}{TODO}

\item{osmext_verbose}{Boolean. If \code{TRUE} the function prints informative messages.}
}
\value{
An sf object related to the input place.
}
\description{
Download, translate and read OSM extracts
}
\details{
This function is a wrapper around ...
}
\examples{
1 + 1
}
