% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectortranslate.R
\name{osmext_vectortranslate}
\alias{osmext_vectortranslate}
\title{Translate the .osm.pbf format into .gpkg}
\usage{
osmext_vectortranslate(
  file_path,
  vectortranslate_options = NULL,
  osmconf_ini = NULL,
  force_vectortranslate = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{file_path}{A}

\item{vectortranslate_options}{B}

\item{osmconf_ini}{C}

\item{force_vectortranslate}{D}

\item{verbose}{E}
}
\value{
path
}
\description{
Translate the .osm.pbf format into .gpkg
}
\examples{
1 + 1
}
