% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmextract-package.R
\docType{package}
\name{osmextract-package}
\alias{osmextract}
\alias{osmextract-package}
\title{osmextract: Download and Import Open Street Map Data Extracts}
\description{
\if{html}{\figure{logo.svg}{options: style='float: right' alt='logo' width='120'}}

Match, download, convert and import Open Street Map data extracts obtained from several providers.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://docs.ropensci.org/osmextract/}
  \item \url{https://github.com/ropensci/osmextract}
  \item Report bugs at \url{https://github.com/ropensci/osmextract/issues}
}

}
\author{
\strong{Maintainer}: Andrea Gilardi \email{andrea.gilardi@unimib.it} (\href{https://orcid.org/0000-0002-9424-7439}{ORCID})

Authors:
\itemize{
  \item Andrea Gilardi \email{andrea.gilardi@unimib.it} (\href{https://orcid.org/0000-0002-9424-7439}{ORCID})
  \item Robin Lovelace (\href{https://orcid.org/0000-0001-5679-6536}{ORCID})
}

Other contributors:
\itemize{
  \item Barry Rowlingson (\href{https://orcid.org/0000-0002-8586-6625}{ORCID}) [contributor]
  \item Salva Fernández (Salva reviewed the package (v. 0.1) for rOpenSci, see <https://github.com/ropensci/software-review/issues/395>) [reviewer]
  \item Nicholas Potter (Nicholas reviewed the package (v. 0.1) for rOpenSci, see <https://github.com/ropensci/software-review/issues/395>) [reviewer]
}

}
\keyword{internal}
