% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_pbf.R
\name{read_pbf}
\alias{read_pbf}
\title{Read pbf files with additional attributes}
\usage{
read_pbf(dsn, layer = "lines",
  attributes = make_additional_attributes(layer = layer),
  ini_file = NULL, append = TRUE)
}
\arguments{
\item{dsn}{The location of the file}

\item{layer}{Which layer to read in? One of
"points"           "lines"            "multipolygons"    "multilinestrings" or "other_relations"}

\item{attributes}{Vector of character strings naming attributes to import}

\item{ini_file}{A modified version of https://github.com/OSGeo/gdal/raw/master/gdal/data/osmconf.ini}

\item{append}{Should the columns named in `attributes` be appended to the default columns?
`TRUE` by default.}
}
\description{
Read pbf files with additional attributes
}
\examples{
\donttest{
pbf_url = geofabric_zones$pbf_url[geofabric_zones$name == "Isle of Wight"]
f = file.path(tempdir(), "test.osm.pbf")
download.file(pbf_url, f)
res = read_pbf(f)
names(res)
res = read_pbf(f, layer = "points")
names(res)
}
}
