% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_zones}
\alias{test_zones}
\title{An sf object of geographical zones taken from download.openstreetmap.fr}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 2 rows and 7 columns.
}
\usage{
test_zones
}
\description{
This object represent a minimal provider's database and it should be used
only for examples and tests.
}
\keyword{datasets}
