% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bbbike_zones}
\alias{bbbike_zones}
\title{A \code{data.frame} of geographical zones taken from bbbike.org}
\format{
A sf dataframe with 430 rows and 14 columns:
\describe{
\item{name}{The, usually English, long-form name of the city}
\item{last_modified}{When was it last modified?}
\item{type}{empty}
\item{pbf_size_size}{Size of the pbf file in bytes.}
\item{base_url}{The base URL for the city}
\item{poly_url}{The .poly file location}
\item{pbf}{Link to the latest .osm.pbf file for this region.}
\item{geometry}{The sfc for that geographical region, rectangular.}
}
}
\source{
\url{https://download.bbbike.org/osm/}
}
\usage{
bbbike_zones
}
\description{
Start bicycle routing for... everywhere!
}
\details{
A \code{sf} \code{data.frame} containing the URLs, names and file_size of the OSM extracts.
See \url{https://download.bbbike.org/osm/}.
}
\keyword{datasets}
