% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geofabrik_zones}
\alias{geofabrik_zones}
\title{sf data frame containing attributes and geometries of geofrabric download zones}
\description{
sf data frame containing attributes and geometries of geofrabric download zones
}
\note{
The `geofabrik_zones` can be (re-)generated using the script in the
`data-raw` directory.
}
\keyword{datasets}
