% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_default_osmconf_ini}
\alias{get_default_osmconf_ini}
\title{Get the path of the osmconf.ini file used by default}
\usage{
get_default_osmconf_ini()
}
\value{
A character vector
}
\description{
Get the path of the osmconf.ini file used by default
}
\examples{
get_default_osmconf_ini()
}
