% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_geofabric.R
\name{get_geofabric}
\alias{get_geofabric}
\title{Download osm data from geofabric}
\usage{
get_geofabric(name = "west-yorkshire", layer = "lines",
  attributes = make_additional_attributes(layer = layer),
  download_directory = gf_download_directory(), ask = TRUE,
  max_dist = 3)
}
\arguments{
\item{name}{Name of the geofabric zone to download}

\item{layer}{Character string telling `sf` which OSM layer to import.
One of `points`, `lines` (the default), `multilinestrings`, `multipolygons` or `other_relations`}

\item{attributes}{Vector of character strings naming attributes to import}

\item{download_directory}{Where to download the data? `tempdir()` by default.}

\item{ask}{Should the user be asked before downloading the file?}

\item{max_dist}{What is the maximum distance in fuzzy matching to tolerate before asking
the user to select which zone to download?}
}
\description{
Download osm data from geofabric
}
\examples{
get_geofabric("isle of man")
\donttest{
get_geofabric(name = "andorra") # try other names, e.g. name = "west-yorkshire"
# user asked to choose closest match when interactive
get_geofabric("kdljfdl", ask = FALSE)
# get zone associated with a point
name = sf::st_sfc(sf::st_point(c(0, 53)), crs = 4326)
get_geofabric(name)
}
}
