% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_geofabric.R
\name{get_geofabric}
\alias{get_geofabric}
\title{Download osm data from geofabric}
\usage{
get_geofabric(name = "west-yorkshire", format = "pbf",
  layer = "lines", download_directory = tempdir())
}
\arguments{
\item{name}{Name of the geofabric zone to download}

\item{format}{Format of data to download (currently only pbf supported)}

\item{layer}{Character string telling `sf` which OSM layer to import.
One of `points`, `lines` (the default), `multilinestrings`, `multipolygons` or `other_relations`}

\item{download_directory}{Where to download the data? `tempdir()` by default.}
}
\description{
Download osm data from geofabric
}
\examples{
get_geofabric("isle of man")
\donttest{
get_geofabric("andorra")
get_geofabric("west-yorkshire")
}
}
