% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gf_find.R
\name{gf_find}
\alias{gf_find}
\title{Find geofabrik zones}
\usage{
gf_find(name, ask = FALSE, max_dist = 5)
}
\arguments{
\item{name}{String or \code{\link[sf:sf]{sf::sf}} spatial object of the geofabrik zone
to download. See examples.}

\item{ask}{Should the user be asked before downloading the file?}

\item{max_dist}{What is the maximum distance in fuzzy matching to tolerate before asking
the user to select which zone to download?}
}
\value{
A data frame representing the matching items from the Geofabrik website
}
\description{
Find geofabrik zones
}
\examples{
gf_find("Andorra")
}
