% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gf_find.R
\name{gf_find_sf}
\alias{gf_find_sf}
\title{Find geofabric zones based on sf or sfc object}
\usage{
gf_find_sf(name, ask = FALSE, op = sf::st_contains)
}
\arguments{
\item{name}{Name of the geofabric zone to download}

\item{ask}{Should the user be asked before downloading the file?}
}
\value{
A data frame representing the matching items from the Geofabrik website
}
\description{
Find geofabric zones based on sf or sfc object
}
\examples{
name = sf::st_sfc(sf::st_point(c(0, 53)), crs = 4326)
gf_find_sf(name)
}
