% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-network-graph.R
\name{net_2_sfnet_undirected}
\alias{net_2_sfnet_undirected}
\title{Convert a spatial network to an undirected sfnetwork}
\usage{
net_2_sfnet_undirected(net_sf)
}
\arguments{
\item{net_sf}{a \code{sf} object representing a spatial network}
}
\value{
a \code{sfnetwork} object
}
\description{
Convert a spatial network to an undirected sfnetwork
}
\examples{
\dontrun{
net_sf <- oe_get_tidynetwork(place = "ITS Leeds", mode = "driving")

sfnet_undirected <- net_2_sfnet_undirected(net_sf)
}
}
