% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match.R
\name{oe_check_pattern}
\alias{oe_check_pattern}
\title{Check for patterns in the provider's data columns}
\usage{
oe_check_pattern(
  pattern,
  provider = "geofabrik",
  match_by = "name",
  full_row = FALSE
)
}
\arguments{
\item{pattern}{Character string for the pattern that should be matched}

\item{provider}{Which provider should be used to download the data? Available
providers can be found with the following command: \code{oe_providers()}.}

\item{match_by}{Column name in the provider dataset used to find the match,
\code{name} by default.}

\item{full_row}{Return all columns in the matching row?
\code{FALSE} by default.}
}
\value{
A
}
\description{
This function is used to explore the provider's data and check for patterns
in the existing columns
}
\examples{
oe_check_pattern(
  pattern = "Yorkshire",
  provider = "geofabrik",
  match_by = "name"
)
res = oe_check_pattern(
  pattern = "Yorkshire",
  provider = "geofabrik",
  match_by = "name",
  full_row = TRUE
)
sf::st_drop_geometry(res)[1:3]
}
