% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{oe_clean}
\alias{oe_clean}
\title{Clean download directory}
\usage{
oe_clean(download_directory = oe_download_directory())
}
\arguments{
\item{download_directory}{The directory where the \code{.osm.pbf} and \code{.gpkg}
files are saved. Default value is \code{oe_download_directory()}.}
}
\value{
The same as \code{unlink()}.
}
\description{
This functions is a wrapper around \code{unlink()} that can be used to delete all
\code{.osm.pbf} and \code{.gpkg} files in a given directory.
}
\examples{
# Warning: the following removes all files in oe_download_directory()
\dontrun{
oe_clean()}
}
