% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{oe_clean}
\alias{oe_clean}
\title{Clean download directory}
\usage{
oe_clean()
}
\value{
The same as \code{unlink()}
}
\description{
This functions can be used to delete all \code{.osm.pbf} and \code{.gpkg} files in the
\code{oe_download_directory}.
}
\examples{
# Warning: the following removes all files in download dir
\donttest{
oe_clean()}
}
