% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{oe_download}
\alias{oe_download}
\title{Download file containing OSM data}
\usage{
oe_download(
  file_url,
  file_basename = basename(file_url),
  provider = infer_provider_from_url(file_url),
  download_directory = oe_download_directory(),
  file_size = NA,
  force_download = FALSE,
  max_file_size = 5e+08,
  quiet = TRUE
)
}
\arguments{
\item{file_url}{A URL containing OSM data (e.g. as a .pbf file)}

\item{file_basename}{The base name of the file. Default behaviour:
auto generated from the URL.}

\item{provider}{Which provider should be used to download the data? Available
providers can be found with the following command: \code{oe_providers()}.}

\item{download_directory}{Where to download the file containing the OSM data?
By default this is \code{tempdir()}, which changes each time you restart R.
You can set a persistent \code{download_directory} by adding the following
to your \code{.Renviron} file (e.g. with \code{usethis::edit_r_environ()}):
\verb{OSMEXT_DOWNLOAD_DIRECTORY=/path/to/osm/data}.}

\item{file_size}{How big is the file? Optional. \code{NA} by default. If it's
bigger than \code{max_file_size} and the function is run in interactive mode,
then an interactive menu is displayed, asking for permission for
downloading the file.}

\item{force_download}{Should the file be updated if it has already been
downloaded? \code{FALSE} by default.}

\item{max_file_size}{The maximum file size to download without asking in
interactive mode. Default: \code{5e+8}, half a gigabyte.}

\item{quiet}{Boolean. If \code{FALSE} the function prints informative messages.}
}
\value{
Character string representing the full path the downloaded file
}
\description{
Download the input file if it's not already present in \code{oe_download_directory()}
or the specified \code{download_directory}.
}
\examples{
its_match = oe_match("ITS Leeds", provider = "test")
oe_download(
  file_url = its_match$url,
  file_size = its_match$file_size,
  # for the moment we always need to specify the provider for files
  # matched with test provider
  provider = "test"
)
\dontrun{
iow_details = oe_match("Isle of Wight")
oe_download(
  file_url = iow_details$url,
  file_size = iow_details$file_size
)
Sucre_details = oe_match("Sucre", provider = "bbbike")
oe_download(
  file_url = Sucre_details$url,
  file_size = Sucre_details$file_size,
  download_directory = tempdir()
)
}
}
