% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{oe_download}
\alias{oe_download}
\title{Download file containing OSM data}
\usage{
oe_download(
  file_url,
  file_basename = basename(file_url),
  provider = infer_provider_from_url(file_url),
  download_directory = oe_download_directory(),
  file_size = NA,
  force_download = FALSE,
  max_file_size = 5e+08,
  verbose = FALSE,
  quiet = FALSE
)
}
\arguments{
\item{file_url}{A URL containing OSM data (e.g. as a .pbf file)}

\item{file_basename}{The base name of the file. Default behaviour:
auto generated from the URL.}

\item{provider}{Which provider should be used to download the data? For the
moment we support only \href{http://download.geofabrik.de/}{\code{"geofabrik"}}.}

\item{download_directory}{TODO}

\item{file_size}{How big is the file? Optional. \code{NA} by default.}

\item{force_download}{TODO}

\item{max_file_size}{TODO}

\item{verbose}{Print information about the file matched? Default: \code{FALSE}.}

\item{quiet}{Should files be downloaded without a progress bar?
\code{FALSE} by default.}
}
\value{
Character string representing the full path the downloaded file
}
\description{
Download the input file if it's not already present in \code{oe_download_directory()}
or the specified \code{download_directory}.
}
\examples{
\dontrun{
iow_details = oe_match("Isle of Wight", provider = "test")
f = oe_download(
  file_url = iow_details$url,
  file_size = iow_details$file_size
)
f
bristol_details = oe_match("Bristol", provider = "bbbike")
oe_download(
  file_url = bristol_details$url,
  file_size = bristol_details$file_size,
  provider = "bbbike",
  download_directory = tempdir()
)
}
}
