% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{oe_download_directory}
\alias{oe_download_directory}
\title{Return the download directory used by the package}
\usage{
oe_download_directory()
}
\value{
A character vector representing the path for the download directory
used by the package.
}
\description{
By default, the download directory is equal to \code{tempdir()}. You can set a
persistent download directory by adding the following command to your
\code{.Renviron} file (e.g. with \code{\link[usethis:edit]{usethis::edit_r_environ()}}):
\verb{OSMEXT_DOWNLOAD_DIRECTORY=/path/to/osm/data}.
}
\examples{
oe_download_directory()
}
