% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{oe_download_directory}
\alias{oe_download_directory}
\title{Return the download directory used by the package}
\usage{
oe_download_directory()
}
\value{
A character vector representing the path for the download directory
used by the package.
}
\description{
Return the download directory used by the package
}
\examples{
oe_download_directory()
}
