% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{oe_download_directory}
\alias{oe_download_directory}
\title{Return the download directory used by the package}
\usage{
oe_download_directory()
}
\value{
A character vector representing the path for the download directory
used by the package.
}
\description{
By default, the download directory is equal to \code{tools::R_user_dir("osmextract", "data")}.
You can set a different persistent or temporary download directory by adding
the following command to your \code{.Renviron} file (e.g. with \code{edit_r_environ}
function in \code{usethis} package): \verb{OSMEXT_DOWNLOAD_DIRECTORY=/path/where/to/save/osm/data}.
}
\examples{
oe_download_directory()
}
