% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-network-graph.R
\name{oe_get_dodgrnetwork}
\alias{oe_get_dodgrnetwork}
\title{Obtain a weighted_streetnet from OpenStreetMap data}
\usage{
oe_get_dodgrnetwork(
  place,
  mode = c("cycling", "driving", "walking"),
  ...,
  highway_filter = NULL,
  quiet = FALSE
)
}
\arguments{
\item{place}{Description of the geographical area that should be matched with
a \code{.osm.pbf} file. Can be either a length-1 character vector, an
\code{sf}/\code{sfc}/\code{bbox} object with any CRS, or a numeric vector of coordinates
with length 2. In the last case, it is assumed that the EPSG code is 4326
specified as c(LON, LAT), while you can use any CRS with \code{sf}/\code{sfc}/\code{bbox}
objects. See Details and Examples in \code{\link[=oe_match]{oe_match()}}.}

\item{mode}{A character string of length one denoting the desired mode of
transport. Can be abbreviated. Currently \code{cycling} (the default), \code{driving}
and \code{walking} are supported.}

\item{...}{additional parameters passed to \code{oe_get()} and \code{dodgr::weight_streetnet()}, excluding \code{x} and \code{id_col} for the latter.}

\item{highway_filter}{string vector of highway types to keep. By default, it includes "motorway", "trunk", "primary", "secondary", "tertiary", "unclassified", and "residential".}

\item{quiet}{Boolean. If \code{FALSE}, the function prints informative messages.
Starting from \code{sf} version
\href{https://r-spatial.github.io/sf/news/index.html#version-0-9-6-2020-09-13}{0.9.6},
if \code{quiet} is equal to \code{FALSE}, then vectortranslate operations will
display a progress bar.}
}
\value{
a \code{dodgr_streetnet} object
}
\description{
This function is a wrapper around \code{oe_get_network()}
that returns a \verb{dodgr_streetnet`` object. It performs simplification of the highway values, filters by highway types and standardises the }oneway\verb{attribute as well as applying the implied oneway restriction based on the}junction`` tag values.
}
\examples{
 highway_filter = c(
 "motorway",
 "trunk",
 "primary",
 "secondary",
 "tertiary",
 "unclassified",
 "residential"
)

 graph_car <- oe_get_dodgrnetwork(
   place = "ITS Leeds",
   mode = "driving",
   wt_profile = "motorcar",
   left_side = TRUE
 )

 class(graph_car)

 graph_bike <- oe_get_dodgrnetwork(
   place = "ITS Leeds",
   mode = "cycling",
   wt_profile = "bicycle",
   left_side = TRUE,
   highway_filter = highway_filter
 )

}
\seealso{
\code{\link[=oe_get]{oe_get()}}, \code{\link[=oe_get_network]{oe_get_network()}}, \code{\link[=oe_get_sfnetwork]{oe_get_sfnetwork()}}, \code{\link[=oe_get_sfnetwork]{oe_get_sfnetwork()}}
}
