% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-network-graph.R
\name{oe_get_dodgrnetwork}
\alias{oe_get_dodgrnetwork}
\title{Obtain a weighted_streetnet object from OpenStreetMap data}
\usage{
oe_get_dodgrnetwork(
  place,
  mode = c("cycling", "driving", "walking"),
  ...,
  highway_filter = NULL,
  quiet = FALSE
)
}
\arguments{
\item{place}{Description of the geographical area that should be matched with
a \code{.osm.pbf} file. Can be either a length-1 character vector, an
\code{sf}/\code{sfc}/\code{bbox} object with any CRS, or a numeric vector of coordinates
with length 2. In the last case, it is assumed that the EPSG code is 4326
specified as c(LON, LAT), while you can use any CRS with \code{sf}/\code{sfc}/\code{bbox}
objects. See Details and Examples in \code{\link[=oe_match]{oe_match()}}.}

\item{mode}{A character string of length one denoting the desired mode of
transport. Can be abbreviated. Currently \code{cycling} (the default), \code{driving}
and \code{walking} are supported.}

\item{...}{additional parameters passed to \code{\link[=oe_get_network]{oe_get_network()}} and
\code{\link[dodgr:weight_streetnet]{dodgr::weight_streetnet()}}, excluding \code{x} and \code{id_col} for the latter.}

\item{highway_filter}{Character vector of highway types to keep.  Ignored if
\code{NULL}. Valid values are: "busway", "cycleway", "footway", "living_street",
"motorway", "path", "pedestrian", "primary", "residential", "rest_area",
"service", "services", "steps", "tertiary", "track", "trunk" and
"unclassified".}

\item{quiet}{Boolean. If \code{FALSE}, the function prints informative messages.
Starting from \code{sf} version
\href{https://r-spatial.github.io/sf/news/index.html#version-0-9-6-2020-09-13}{0.9.6},
if \code{quiet} is equal to \code{FALSE}, then vectortranslate operations will
display a progress bar.}
}
\value{
A \code{dodgr_streetnet} object
}
\description{
This function is a wrapper around \code{oe_get_network()} that returns a
\code{dodgr_streetnet} object. It performs simplification of the highway values,
filters by highway types and standardises the \code{oneway} attribute as well as
applying the implied oneway restriction based on the `junction`` tag values.
}
\examples{
# Copy the ITS file to tempdir() to make sure that the examples do not
# require internet connection. You can skip the next 4 lines (and start
# directly with oe_get_keys) when running the examples locally.
its_pbf = file.path(tempdir(), "test_its-example.osm.pbf")
file.copy(
  from = system.file("its-example.osm.pbf", package = "osmextract"),
  to = its_pbf,
  overwrite = TRUE
)

graph_bike <- oe_get_dodgrnetwork(
  place = "ITS Leeds",
  mode = "cycling",
  wt_profile = "bicycle",
  left_side = TRUE,
  download_directory = tempdir()
)
head(graph_bike)

highway_filter = c(
  "motorway",
  "trunk",
  "primary",
  "secondary",
  "tertiary",
  "unclassified",
  "residential"
)

 graph_car <- oe_get_dodgrnetwork(
   place = "ITS Leeds",
   mode = "driving",
   wt_profile = "motorcar",
   left_side = TRUE,
   highway_filter = highway_filter,
   download_directory = tempdir(),
   quiet = TRUE
 )

 head(graph_car)
 class(graph_car)
}
\seealso{
\code{\link[=oe_get]{oe_get()}}, \code{\link[=oe_get_network]{oe_get_network()}}, \code{\link[=oe_get_sfnetwork]{oe_get_sfnetwork()}}
}
