% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-key-values.R
\name{oe_get_keys}
\alias{oe_get_keys}
\alias{oe_get_keys.default}
\alias{oe_get_keys.character}
\alias{oe_get_keys.sf}
\alias{print.oe_key_values_list}
\title{Return keys and (optionally) values stored in "other_tags" column}
\usage{
oe_get_keys(zone, layer = "lines", values = FALSE, which_keys = NULL)

\method{oe_get_keys}{default}(zone, layer = "lines", values = FALSE, which_keys = NULL)

\method{oe_get_keys}{character}(zone, layer = "lines", values = FALSE, which_keys = NULL)

\method{oe_get_keys}{sf}(zone, layer = "lines", values = FALSE, which_keys = NULL)

\method{print}{oe_key_values_list}(x, n = getOption("oe_max_print_keys", 10L), ...)
}
\arguments{
\item{zone}{An \code{sf} object with an \code{other_tags} field or a character vector
(of length 1) pointing to a \code{.osm.pbf} or \code{.gpkg} file with an \code{other_tags}
field.}

\item{layer}{Which \code{layer} should be read in? Typically \code{points}, \code{lines}
(the default), \code{multilinestrings}, \code{multipolygons} or \code{other_relations}. If
you specify an ad-hoc query using the argument \code{query} (see introductory
vignette and examples), then \code{oe_get()} and \code{oe_read()} will read the layer
specified in the query and ignore \code{layer}. See also
\href{https://github.com/ropensci/osmextract/issues/122}{#122}.}

\item{values}{Logical. If \code{TRUE}, then function returns the keys and the
corresponding values, otherwise only the keys. Defaults to \code{FALSE. }}

\item{which_keys}{Subset only some keys and corresponding values. Ignored if
\code{values} is \code{FALSE}. See examples.}

\item{x}{object of class \code{oe_key_values_list}}

\item{n}{Maximum number of keys (and corresponding values) to print; can be
set globally by \code{options(oe_max_print_keys=...)}. Default value is 10.}

\item{...}{Ignored.}
}
\value{
If \code{values} is \code{FALSE} (the default), then the function returns a
character vector with the names of all keys stored in the "other_tags"
field. If \code{values} is \code{TRUE}, then the function returns named list which
stores all keys and the corresponding values. In the latter case, the
returned object has class \code{oe_key_values_list} and we defined an ad-hoc
printing method. See Details.
}
\description{
This function returns the keys and (optionally) the values stored in the
\code{other_tags} field. See Details. In both cases, the keys are sorted according
to the number of occurrences, which means that the most common keys are
reported first.
}
\details{
OSM data are typically documented using several
\href{https://wiki.openstreetmap.org/wiki/Tags}{\code{tags}}, i.e. pairs of two
items, namely a \code{key} and a \code{value}. The conversion between \code{.osm.pbf} and
\code{.gpkg} formats is governed by a \code{CONFIG} file that lists which tags must
be explicitly added to the \code{.gpkg} file. All the other keys are
automatically stored using an \code{other_tags} field with a syntax compatible
with the PostgreSQL HSTORE type. See
\href{https://gdal.org/drivers/vector/osm.html#driver-capabilities}{here} for
more details.

When \code{values} is \code{TRUE}, then the function returns a named list of class
\code{oe_key_values_list} that, for each key, summarises the corresponding
values. The object is stored using the following format:
\verb{list(key1 = c("value1", "value1", "value2", ...), key2 = c("value1", ...) ...)}
Nevertheless, we decided to implement an ad-hoc method for printing objects
of class \code{oe_key_values_list} using the following structure: \preformatted{
`key1 = {#value1 = n1; #value2 = n2; #value3 = n3, ...}`
`key2 = {#value1 = n1; #value2 = n2; ...}`
`key3 = {#value1 = n1}`
}
and so on. By default, the function prints only the ten most common keys,
but the value can be adjusted using the option \code{oe_max_print_keys}.

Finally, the \code{hstore_get_value()} function can be used inside the \code{query}
argument in \code{oe_get()} to extract one particular tag from an existing file.
Check the introductory vignette and see examples.
}
\examples{
# Get keys from pbf file
itsleeds_pbf_path = oe_download(
  oe_match("ITS Leeds")$url,
  download_directory = tempdir(),
  provider = "test"
)
oe_get_keys(itsleeds_pbf_path)
itsleeds_gpkg_path = oe_get(
  "ITS Leeds",
  download_only = TRUE,
  download_directory = tempdir(),
  quiet = TRUE
)
itsleeds_gpkg_path
oe_get_keys(itsleeds_gpkg_path)

itsleeds = oe_get("ITS Leeds", quiet = TRUE, download_directory = tempdir())
oe_get_keys(itsleeds)

# Add an extra key to an existing .gpkg file without vectortranslate
names(oe_read(
  itsleeds_gpkg_path,
  query = "SELECT *,  hstore_get_value(other_tags, 'oneway')  AS oneway FROM lines"
))

# Remove .pbf and .gpkg files in tempdir
# (since they may interact with other examples)
file.remove(list.files(path = tempdir(), pattern = "(pbf|gpkg)", full.names = TRUE))
}
\seealso{
\code{oe_vectortranslate()}
}
