% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectortranslate.R
\name{oe_get_keys}
\alias{oe_get_keys}
\title{Return all keys stored in "other_tags" column}
\usage{
oe_get_keys(file_path, layer = "lines")
}
\arguments{
\item{file_path}{The path of a \code{.gpkg} file, typically created using
\code{oe_vectortranslate()} or \code{oe_get()}.}

\item{layer}{Which \code{layer} should be read in? Typically \code{points}, \code{lines}
(the default), \code{multilinestrings}, \code{multipolygons} or \code{other_relations}.}
}
\value{
A character vector indicating the name of all keys stored in
"other_tags" field.
}
\description{
This function is used to return the names of all keys that are stored in
"other_tags" column since they were not explicitly included in the file. See
Details.
}
\details{
OSM data are typically documented using several
\href{https://wiki.openstreetmap.org/wiki/Tags}{\code{tags}}. A \code{tag} is a pair of
two items, namely a \code{key} and a \code{value}. As we documented in
\code{oe_vectortranslate()}, the conversion between \code{.osm.pbf} and \code{.gpkg}
formats is governed by a CONFIG file that indicates which tags are
explicitly added to the \code{.gpkg} file. All the other keys stored in the
\code{.osm.pbf} file are automatically appended in an "other_tags" field, with a
syntax compatible with the PostgreSQL HSTORE type. This function is used to
display the names of all the keys stored in the "other_tags" column. See
examples.

You can also use the \code{hstore_get_value} function from GDAL to extract one
particular tag from an existing .gpkg file. Check the introductory vignette
for a more detailed explanation and see examples.
}
\examples{
itsleeds_gpkg = oe_get("itsleeds", provider = "test", download_only = TRUE)
oe_get_keys(itsleeds_gpkg)

# Add an extra key to an existing .gpkg file without the .pbf
names(oe_read(itsleeds_gpkg, extra_tags = c("oneway"))) # doesn't work
# we need a different approach
names(oe_read(
  itsleeds_gpkg,
  query = "SELECT *,  hstore_get_value(other_tags, 'oneway')  AS oneway FROM lines"
))
}
\seealso{
\code{oe_vectortranslate()} and
\href{https://github.com/ITSLeeds/osmextract/issues/107}{#107}.
}
