% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-network.R
\name{oe_get_network}
\alias{oe_get_network}
\title{Import a road network according to a specific mode of transport}
\usage{
oe_get_network(place, mode = c("cycling", "driving", "walking"), ...)
}
\arguments{
\item{place}{Description of the geographical area that should be matched with
a \code{.osm.pbf} file. Can be either a length-1 character vector, an
\code{sf}/\code{sfc}/\code{bbox} object, or a numeric vector of coordinates with length 2.
In the last case, it is assumed that the EPSG code is 4326 specified as
c(LON, LAT), while you can use any CRS with \code{sf}/\code{sfc}/\code{bbox} objects. See
Details and Examples in \code{\link[=oe_match]{oe_match()}}.}

\item{mode}{A character string denoting the desired mode of transport. Can be
abbreviated.}

\item{...}{IGNORED FOR THE MOMENT. Additional arguments passed to \code{oe_get()}.
Please note that the arguments \code{layer} and \code{vectortranslate_options} cannot
be provided since are set according to the \code{mode} parameter.}
}
\value{
An \code{sf} object.
}
\description{
Import a road network according to a specific mode of transport
}
\details{
TODO (precisely describe the function) Also slightly underline the
differnces between legal access and convenience (i.e. these are not the
best paths, just all available paths)
}
\examples{
oe_get_network("ITS Leeds", mode = "walking")
}
\seealso{
\code{\link[=oe_get]{oe_get()}}
}
