% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-network-graph.R
\name{oe_get_sfnetwork}
\alias{oe_get_sfnetwork}
\title{Obtain a sfnetwork object from OpenStreetMap data}
\usage{
oe_get_sfnetwork(
  place,
  mode = c("cycling", "driving", "walking"),
  ...,
  require_equal = TRUE,
  directed = FALSE,
  highway_filter = NULL,
  quiet = FALSE
)
}
\arguments{
\item{place}{Description of the geographical area that should be matched with
a \code{.osm.pbf} file. Can be either a length-1 character vector, an
\code{sf}/\code{sfc}/\code{bbox} object with any CRS, or a numeric vector of coordinates
with length 2. In the last case, it is assumed that the EPSG code is 4326
specified as c(LON, LAT), while you can use any CRS with \code{sf}/\code{sfc}/\code{bbox}
objects. See Details and Examples in \code{\link[=oe_match]{oe_match()}}.}

\item{mode}{A character string of length one denoting the desired mode of
transport. Can be abbreviated. Currently \code{cycling} (the default), \code{driving}
and \code{walking} are supported.}

\item{...}{Additional arguments passed to \code{\link[=oe_get_network]{oe_get_network()}}.}

\item{require_equal}{logical or character vector. Defaults to \code{TRUE.} If
\code{TRUE}, only pseudo nodes that have incident edges with equal attribute
values are removed. Alternatively, a character vector with the names of the
attributes to check for equality can be provided. In that case, only those
attributes are checked for equality. If \code{FALSE}, all pseudo nodes are
removed regardless of attribute values. All other attributes are collapsed,
separating the values by a comma.}

\item{directed}{logical, whether to return a directed sfnetwork object
(default is \code{FALSE})}

\item{highway_filter}{Character vector of highway types to keep. Ignored if clean_output is FALSE.
Valid values are: "busway", "cycleway", "footway", "living_street", "motorway",
"path", "pedestrian", "primary", "residential", "rest_area", "service", "services",
"steps", "tertiary", "track", "trunk" and "unclassified".}

\item{quiet}{Boolean. If \code{FALSE}, the function prints informative messages.
Starting from \code{sf} version
\href{https://r-spatial.github.io/sf/news/index.html#version-0-9-6-2020-09-13}{0.9.6},
if \code{quiet} is equal to \code{FALSE}, then vectortranslate operations will
display a progress bar.}
}
\value{
An \code{sfnetwork} object
}
\description{
This function is a wrapper around \code{oe_get_network()} that returns a
\code{sfnetwork} object. It performs simplification of the highway values and
filters by highway types if specified. Minimal network preprocessing tasks
i.e. subdivision and smoothing are performed. It also allows for the creation
of directed or undirected networks.
}
\details{
Note that preprocessing is performed on the undirected graph, ignoring road
directionality. See the documentation for \code{sfnetworks} for more details on
the two preprocessing steps: \href{https://luukvdmeer.github.io/sfnetworks/articles/sfn02_preprocess_clean.html#subdivide-edges}{subdividing edges}
and \href{https://luukvdmeer.github.io/sfnetworks/articles/sfn02_preprocess_clean.html#smooth-pseudo-nodes}{smoothing pseudo-nodes}.
If \code{directed = TRUE}, the function then builds the directed graph by
duplicating bidirectional edges with reversed geometries. If the \code{oneway}
column is missing, a warning is issued and an undirected \code{sfnetwork} is
returned instead.
}
\examples{
# Copy the ITS file to tempdir() to make sure that the examples do not
# require internet connection. You can skip the next 4 lines (and start
# directly with oe_get_keys) when running the examples locally.
its_pbf = file.path(tempdir(), "test_its-example.osm.pbf")
file.copy(
  from = system.file("its-example.osm.pbf", package = "osmextract"),
  to = its_pbf,
  overwrite = TRUE
)
# Undirected unfiltered pedestrian network
walk_sfnet <- oe_get_sfnetwork(
  place = "ITS Leeds",
  mode = "walking",
  download_directory = tempdir(),
  quiet = TRUE
)
plot(walk_sfnet)

# Directed unfiltered driving network
car_sfnet <- oe_get_sfnetwork(
  place = "ITS Leeds",
  mode = "driving",
  directed = TRUE,
  download_directory = tempdir(),
  quiet = TRUE
)
plot(car_sfnet)

# Directed filtered driving network
highway_filter = c(
 "motorway",
 "trunk",
 "primary",
 "secondary",
 "tertiary",
 "unclassified",
 "residential"
 )

car_sfnet_filtered <- oe_get_sfnetwork(
  place = "ITS Leeds",
  mode = "driving",
  directed = TRUE,
  highway_filter = highway_filter,
  download_directory = tempdir(),
  quiet = TRUE
)
plot(car_sfnet_filtered)

# Smooth pseudo nodes when highway and oneway fields are equal
car_sfnet_filtered_2 <- oe_get_sfnetwork(
  place = "ITS Leeds",
  mode = "driving",
  directed = TRUE,
  highway_filter = highway_filter,
  require_equal = c("highway", "oneway"),
  download_directory = tempdir(),
  quiet = TRUE
)
plot(car_sfnet_filtered_2)
}
\seealso{
\code{\link[=oe_get]{oe_get()}}, \code{\link[=oe_get_network]{oe_get_network()}}, \code{\link[=oe_get_dodgrnetwork]{oe_get_dodgrnetwork()}}
}
