% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-network-graph.R
\name{oe_get_sfnetwork}
\alias{oe_get_sfnetwork}
\title{Obtain a sfnetwork object from OpenStreetMap data}
\usage{
oe_get_sfnetwork(
  place,
  mode = c("cycling", "driving", "walking"),
  ...,
  directed = FALSE,
  highway_filter = NULL,
  quiet = FALSE
)
}
\arguments{
\item{place}{Description of the geographical area that should be matched with
a \code{.osm.pbf} file. Can be either a length-1 character vector, an
\code{sf}/\code{sfc}/\code{bbox} object with any CRS, or a numeric vector of coordinates
with length 2. In the last case, it is assumed that the EPSG code is 4326
specified as c(LON, LAT), while you can use any CRS with \code{sf}/\code{sfc}/\code{bbox}
objects. See Details and Examples in \code{\link[=oe_match]{oe_match()}}.}

\item{mode}{A character string of length one denoting the desired mode of
transport. Can be abbreviated. Currently \code{cycling} (the default), \code{driving}
and \code{walking} are supported.}

\item{...}{Additional arguments passed to \code{\link[=oe_get]{oe_get()}} such as \code{boundary} or
\code{force_download}.}

\item{directed}{logical, whether to return a directed sfnetwork object (default is FALSE)}

\item{highway_filter}{Character vector of highway types to keep. Ignored if clean_output is FALSE.
Valid values are: "busway", "cycleway", "footway", "living_street", "motorway",
"path", "pedestrian", "primary", "residential", "rest_area", "service", "services",
"steps", "tertiary", "track", "trunk" and "unclassified".}

\item{quiet}{Boolean. If \code{FALSE}, the function prints informative messages.
Starting from \code{sf} version
\href{https://r-spatial.github.io/sf/news/index.html#version-0-9-6-2020-09-13}{0.9.6},
if \code{quiet} is equal to \code{FALSE}, then vectortranslate operations will
display a progress bar.}
}
\value{
An \code{sfnetwork} object
}
\description{
This function is a wrapper around \code{oe_get_network()} that returns a \code{sfnetwork} object.
It performs simplification of the highway values and filters by highway types if specified. Minimal
network preprocessing tasks i.e. subdivision and smoothing are performed using \code{sfnetworks} to
create a clean \code{sfnetwork} object. All unique merged edge attributes are concatenated.
It also allows for the creation of directed or undirected networks.
}
\examples{
highway_filter = c(
 "motorway",
 "trunk",
 "primary",
 "secondary",
 "tertiary",
 "unclassified",
 "residential"
)

# sfnet directed unfiltered
car_sfnet <- oe_get_sfnetwork(
  place = "ITS Leeds",
  mode = "driving",
  directed = TRUE
)
plot(car_sfnet)
# sfnet directed filtered
car_sfnet_filtered <- oe_get_sfnetwork(
  place = "ITS Leeds",
  mode = "driving",
  directed = TRUE,
  highway_filter = highway_filter
)

# sfnet_undirected filtered
walk_sfnet <- oe_get_sfnetwork(
  place = "ITS Leeds",
  mode = "walking"
)
plot(walk_sfnet)

}
\seealso{
\code{\link[=oe_get]{oe_get()}}, \code{\link[=oe_get_network]{oe_get_network()}}, \code{\link[=oe_get_dodgrnetwork]{oe_get_dodgrnetwork()}}
}
