% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-network-graph.R
\name{oe_get_sfnetwork}
\alias{oe_get_sfnetwork}
\title{Obtain a sfnetwork object from OpenStreetMap data}
\usage{
oe_get_sfnetwork(
  ...,
  directed = FALSE,
  simplify_highway = TRUE,
  highway_filter = NULL
)
}
\arguments{
\item{...}{parameters passed to \code{oe_get_network()}}

\item{directed}{logical, whether to return a directed sfnetwork object (default is FALSE)}

\item{simplify_highway}{logical, whether to simplify the highway values by removing the "_link" suffix and filtering by \code{highway_filter}}

\item{highway_filter}{character vector of highway types to keep, if \code{simplify_highway} is TRUE}
}
\value{
a \code{sfnetwork} object
}
\description{
This function is a wrapper around \code{oe_get_network()} that returns a \code{sfnetwork} object.
It performs simplification of the highway values and filters by highway types if specified. Minimal
network preprocessing tasks i.e. subdivision and smoothing are performed using \code{sfnetworks} to
create a tidy \code{sfnetwork} object. All unique merged edge attributes are concatenated.
It also allows for the creation of directed or undirected networks.
}
\examples{
highway_filter = c(
 "motorway",
 "trunk",
 "primary",
 "secondary",
 "tertiary",
 "unclassified",
 "residential"
)

# sfnet directed unfiltered
car_sfnet <- oe_get_sfnetwork(
  place = "ITS Leeds",
  mode = "driving",
  directed = TRUE
)
plot(car_sfnet)
# sfnet directed filtered
car_sfnet_filtered <- oe_get_sfnetwork(
  place = "ITS Leeds",
  mode = "driving",
  directed = TRUE,
  highway_filter = highway_filter
)

# sfnet_undirected filtered
walk_sfnet <- oe_get_sfnetwork(
  place = "ITS Leeds",
  mode = "walking"
)
plot(walk_sfnet)

}
\seealso{
\code{\link[=oe_get]{oe_get()}}, \code{\link[=oe_get_network]{oe_get_network()}}, \code{\link[=oe_get_tidynetwork]{oe_get_tidynetwork()}}, \code{\link[=oe_get_dodgrnetwork]{oe_get_dodgrnetwork()}}
}
