% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-network-graph.R
\name{oe_get_tidynetwork}
\alias{oe_get_tidynetwork}
\title{Obtain a tidy sf from OpenStreetMap data}
\usage{
oe_get_tidynetwork(..., simplify_highway = TRUE, highway_filter = NULL)
}
\arguments{
\item{...}{parameters passed to \code{oe_get_network()}}

\item{simplify_highway}{logical, whether to simplify the highway values by removing the "_link" suffix and filtering by \code{highway_filter}}

\item{highway_filter}{character vector of highway types to keep, if \code{simplify_highway} is TRUE}
}
\value{
a \code{sf} object with standardised highway and oneway values
}
\description{
Obtain a tidy sf from OpenStreetMap data
}
\examples{

highway_filter = c(
 "motorway",
 "trunk",
 "primary",
 "secondary",
 "tertiary",
 "unclassified",
 "residential"
)

tidynet_sf <- oe_get_tidynetwork(place = "ITS Leeds", mode = "driving")
print(tidynet_sf)

tidynet_sf_filtered <- oe_get_tidynetwork(
  place = "ITS Leeds",
  mode = "driving",
  highway_filter = highway_filter
)
print(tidynet_sf_filtered)

}
