% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{oe_read}
\alias{oe_read}
\title{Read a .pbf or .gpkg object}
\usage{
oe_read(
  file_path,
  layer = "lines",
  ...,
  download_directory = oe_download_directory(),
  vectortranslate_options = NULL,
  osmconf_ini = NULL,
  extra_attributes = NULL,
  force_vectortranslate = NULL,
  skip_vectortranslate = FALSE,
  quiet = TRUE
)
}
\arguments{
\item{file_path}{The path of the .gpkg or .osm.pbf file that should be
read-in.}

\item{layer}{Which \code{layer} should be read in? Typically \code{points}, \code{lines}
(the default), \code{multilinestrings}, \code{multipolygons} or \code{other_relations}.}

\item{...}{Arguments that should  be passed to \code{\link[sf:st_read]{sf::st_read()}}}

\item{download_directory}{Where to download the file containing the OSM data?
By default this is \code{tempdir()}, which changes each time you restart R.
You can set a persistent \code{download_directory} by adding the following
to your \code{.Renviron} file (e.g. with \code{usethis::edit_r_environ()}):
\verb{OSMEXT_DOWNLOAD_DIRECTORY=/path/to/osm/data}.}

\item{vectortranslate_options}{Options to pass to the \code{\link[sf:gdal_utils]{sf::gdal_utils()}}
argument \code{options}. Set by default. Check Details at
\code{oe_vectortranslate()}.}

\item{osmconf_ini}{The configuration file specifying which columns should be
in the resulting data frame. See documentation at
\href{https://gdal.org/drivers/vector/osm.html}{gdal.org}. Check Details at
\code{oe_vectortranslate()}.}

\item{extra_attributes}{Which addition columns, corresponding to OSM keys,
should be in the resulting dataset? \code{FALSE} by default. Check Details at
\code{oe_vectortranslate()} and \code{oe_get_keys()}.}

\item{force_vectortranslate}{Boolean. Force the original \code{.pbf} file to be
translated into a \code{.gpkg} file, even if a \code{.gpkg} with the same name
already exists? Check Details at \code{oe_vectortranslate()}.}

\item{skip_vectortranslate}{Boolean. If \code{TRUE} then the function skips all
the vectortranslate operations and it reads (or simply returns the path) of
the \code{.osm.pbf} file. \code{FALSE} by default.}

\item{quiet}{Boolean. If \code{FALSE} the function prints informative messages.}
}
\value{
An sf object related to the input path.
}
\description{
Read a .pbf or .gpkg object
}
\details{
For the moment I will consider only .gpkg or .osm.pbf file. The
approach is more or less the same as for oe_get() but we need to skip the
matching operations. In the near future I think I could merge the two
approaches i.e. oe_get will be oe_match + oe_read to avoid a lot of
duplication.
}
\examples{
oe_read(
  file_path = system.file("its-example.osm.pbf", package = "osmextract"),
  download_directory = tempdir()
)
}
