% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{oe_update}
\alias{oe_update}
\title{Update all the \code{.osm.pbf} files saved in a directory}
\usage{
oe_update(
  download_directory = oe_download_directory(),
  oe_verbose = TRUE,
  delete_gpkg = TRUE,
  ...
)
}
\arguments{
\item{download_directory}{Character string of the path of the directory
where the \code{.osm.pbf} files are saved.}

\item{oe_verbose}{Boolean. If \code{TRUE} the function prints informative
messages. See Details.}

\item{delete_gpkg}{Boolean. if \code{TRUE} the function deletes the old \code{.gpkg}
files. We added this parameter to minimize the probability of accidentally
reading-in old and not-synchronized \code{.gpkg} files. See details. Defaults to
\code{TRUE}.}

\item{...}{Additional parameter that will be passed to \code{oe_get()}.}
}
\value{
The path(s) of the .osm.pbf file(s) that were updated invisibly.
}
\description{
This function is used to re-download all \code{.osm.pbf} files stored in
\code{download_directory} that were firstly downloaded through \code{oe_get()}. See
details.
}
\details{
This function is used to re-download the \code{.osm.pbf} files stored in
a the directory (specified by \code{download_directory} param) that were firstly
downloaded through \code{oe_get()}. The name of the files must begin with the
name of one of the supported providers (see \code{oe_available_providers()}) and
it must end with \code{".osm.pbf"}. All other files in the directory that do
not match this format are ignored.

The process for re-downloading the \code{.osm.pbf} files is performed using the
function \code{oe_get()}. The appropriate provider is determined by looking at
the first word in the path of the \code{.osm.pbf} file. The place is determined
by looking at the second word in the file path and the matching is
performed through the \code{id} column in the provider's database. So, for
example, the path \code{geofabrik_italy-latest-update.osm.pbf} will be matched
with the provider \code{"geofabrik"} and the geographical zone \code{italy} through
the column \code{id} in \code{geofabrik_zones}.

The parameter \code{delete_gpkg} is used to delete all \code{.gpkg} files in
\code{download_directory}. We decided to set its default value to \code{TRUE} to
minimize the possibility of reading-in old and non-synchonized \code{.gpkg}
files. If you set \code{delete_gpkg = TRUE}, then you need to manually reconvert
all files using \code{oe_get()} or \code{oe_vectortranslate()}. See examples.
}
\examples{
1 + 1
}
