% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectortranslate.R
\name{oe_vectortranslate}
\alias{oe_vectortranslate}
\title{Translate a \code{.osm.pbf} file into \code{.gpkg} format}
\usage{
oe_vectortranslate(
  file_path,
  vectortranslate_options = NULL,
  layer = "lines",
  osmconf_ini = NULL,
  extra_attributes = NULL,
  force_vectortranslate = NULL,
  quiet = TRUE
)
}
\arguments{
\item{file_path}{Character string representing the path of the input
\code{.osm.pbf} file}

\item{vectortranslate_options}{Options to pass to the \code{\link[sf:gdal_utils]{sf::gdal_utils()}}
argument \code{options}. Set by default. Check details and examples.}

\item{layer}{Which \code{layer} should be read in? Typically \code{points}, \code{lines}
(the default), \code{multilinestrings}, \code{multipolygons} or \code{other_relations}.}

\item{osmconf_ini}{The configuration file specifying which columns should be
in the resulting data frame. See documentation at
\href{https://gdal.org/drivers/vector/osm.html}{gdal.org} for details.}

\item{extra_attributes}{Which addition columns, corresponding to OSM keys,
should be in the resulting dataset? \code{NULL} by default.}

\item{force_vectortranslate}{Force the original \code{.pbf} file to be translated
into a \code{.gpkg} file, even if a \code{.gpkg} associated with the \code{provider} zone
already exists.}

\item{quiet}{Boolean. If \code{FALSE} the function prints informative messages.}
}
\value{
Character string representing the path of the .gpkg file.
}
\description{
This function is used to translate a \code{.osm.pbf} file into \code{.gpkg} format. The
conversion is performed using
\href{https://gdal.org/programs/ogr2ogr.html#ogr2ogr}{ogr2ogr} through
\code{sf::gdal_utils()}. It was created following \href{https://github.com/OSGeo/gdal/issues/2100#issuecomment-565707053}{the suggestions}
of the maintainers of GDAL. See Details and Examples to understand the basic
functionalities, and check the introductory vignette for more complex
use-cases.
}
\details{
The new \code{.gpkg} file is created in the same directory as the input
\code{.osm.pbf} file. The translation process is performed using the
\code{sf::gdal_utils()} function, setting \code{util = "vectortranslate"}. This
operation can be customized in several ways modifying the parameters
\code{vectortranslate_options}, \code{layer}, \code{osmconf_ini} and \code{extra_attributes}.

The \code{.osm.pbf} files that are read using GDAL are usually categorized into
5 layers, named \code{points}, \code{lines}, \code{multilinestrings}, \code{multipolygons} and
\code{other_relations}. Check the first paragraphs
\href{https://gdal.org/drivers/vector/osm.html}{here} for more details.  The
parameter \code{layer} is used to specify which layer of the \code{.osm.pbf} file
should be converted into the \code{.gpkg} file. Several layers with different
names can be stored in the same \code{.gpkg} file. We can convert only one layer
at time. By default, the function will convert the \code{lines} layer (which is
the most common one according to our experience). The vectortranslate
operation is skipped if the function detects a file having the same name as
the input file, \code{.gpkg} extension and a layer with the same name as the
parameter \code{layer}. This behaviour can be overwritten by setting
\code{force_vectortranslate = TRUE}.

The arguments \code{osmconf_ini} and \code{extra_attributes} are used to modify how
GDAL reads the \code{.osm.pbf} file. More precisely, the parameter \code{osmconf_ini}
must be a character string specifying the path of the \code{.ini} file used by
GDAL.
}
\examples{
# First we need to match an input zone with a .osm.pbf file
its_match = oe_match("ITS Leeds", provider = "test")
# The we can download the .osm.pbf files
its_pbf = oe_download(
  file_url = its_match$url,
  file_size = its_match$file_size,
  download_directory = tempdir(),
  provider = "test"
)
list.files(tempdir(), pattern = "pbf|gpkg")
its_gpkg = oe_vectortranslate(
 its_pbf
)
list.files(tempdir(), pattern = "pbf|gpkg")
}
