% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectortranslate.R
\name{oe_vectortranslate}
\alias{oe_vectortranslate}
\title{Translate the .osm.pbf format into .gpkg}
\usage{
oe_vectortranslate(
  file_path,
  vectortranslate_options = NULL,
  layer = NULL,
  osmconf_ini = NULL,
  extra_attributes = NULL,
  force_vectortranslate = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{file_path}{A}

\item{vectortranslate_options}{Options to pass to the \code{\link[sf:gdal_utils]{sf::gdal_utils()}}
argument \code{options}. Set by default.}

\item{layer}{Which \code{layer} should be read in? Typically \code{points}, \code{lines}
(the default), \code{multilinestrings}, \code{multipolygons} or \code{other_relations}.}

\item{osmconf_ini}{The configuration file specifying which columns should be
in the resulting data frame. See documentation at
\href{https://gdal.org/drivers/vector/osm.html}{gdal.org} for details.}

\item{extra_attributes}{Which addition columns, corresponding to OSM keys,
should be in the resulting dataset? \code{NULL} by default.}

\item{force_vectortranslate}{Force the original \code{.pbf} file to be translated
into a \code{.gpkg} file, even if a \code{.gpkg} associated with the \code{provider} zone
already exists.}

\item{verbose}{Boolean. If \code{TRUE} the function prints informative messages.}
}
\value{
path
}
\description{
Translate the .osm.pbf format into .gpkg
}
\examples{
1 + 1
}
