% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectortranslate.R
\name{oe_vectortranslate}
\alias{oe_vectortranslate}
\title{Translate the .osm.pbf format into .gpkg}
\usage{
oe_vectortranslate(
  file_path,
  vectortranslate_options = NULL,
  layer = NULL,
  osmconf_ini = NULL,
  extra_attributes = NULL,
  force_vectortranslate = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{file_path}{A}

\item{vectortranslate_options}{B}

\item{layer}{B2}

\item{osmconf_ini}{C}

\item{extra_attributes}{D}

\item{force_vectortranslate}{E}

\item{verbose}{F}
}
\value{
path
}
\description{
Translate the .osm.pbf format into .gpkg
}
\examples{
1 + 1
}
