% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match.R
\name{osmext_check_pattern}
\alias{osmext_check_pattern}
\title{Check for patterns in the provider's data columns}
\usage{
osmext_check_pattern(pattern, provider = "geofabrik", match_by = "name")
}
\arguments{
\item{pattern}{Character string for the pattern that should be matched}

\item{provider}{Which provider should be used to download the data associated
to the mathced \code{place}? For the moment we support only
\href{http://download.geofabrik.de/}{\code{"geofabrik"}}.}

\item{match_by}{Which column of the provider data should be used for matching
the input place? The default value is "name". Check details and examples to
understand why this parameter is important.}
}
\value{
A
}
\description{
This function is used to explore the provider's data and check for patterns
in the existing columns
}
\examples{
osmext_check_pattern(
pattern = "Yorkshire",
provider = "geofabrik",
match_by = "name"
)
}
