% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{osmext_download}
\alias{osmext_download}
\title{Download the input file}
\usage{
osmext_download(
  file_url,
  file_basename = basename(file_url),
  provider = infer_provider_from_url(file_url),
  download_directory = osmext_download_directory(),
  file_size = NA,
  force_download = FALSE,
  max_file_size = 5e+08,
  verbose = FALSE
)
}
\arguments{
\item{file_url}{A}

\item{file_basename}{B}

\item{provider}{Which provider should be used to download the data? For the
moment we support only \href{http://download.geofabrik.de/}{\code{"geofabrik"}}.}

\item{download_directory}{TODO}

\item{file_size}{E}

\item{force_download}{TODO}

\item{max_file_size}{TODO}

\item{verbose}{Boolean. If \code{TRUE} the function prints informative messages.}
}
\value{
path
}
\description{
Download the input file if it's not already present in the specified download_directory.
}
\examples{
1 + 1
}
