% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{osmext_get}
\alias{osmext_get}
\title{Title}
\usage{
osmext_get(
  place,
  provider = "geofabrik",
  match_by = "name",
  max_string_dist = 1,
  interactive_ask = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{place}{Description of the geographical area that should be download
through the chosen \code{provider}. Can be either a length-1 character vector, a
length-1 \code{sfc_POINT} object or a numeric vector with length 2. See details
and examples.}

\item{provider}{Which provider should be used to download the data associated
to the input \code{place}? For the moment we support only
\href{http://download.geofabrik.de/}{\code{"geofabrik"}}.}

\item{match_by}{Which column of the provider data should be used for matching
the input place with the provider's data? The default is "name". Check
details and examples to understand how this parameter works. Ignored if
\code{place} is not a character vector since the matching is performed through a
spatial operation.}

\item{max_string_dist}{Numeric and ge than 0. What is the maximum distance in
fuzzy matching to tolerate before asking the user to select which zone to
download? Check details and examples to understand why this parameter is
important. Ignored if \code{place} is not a character vector since the matching
is performed through a spatial operation.}

\item{interactive_ask}{Boolean. If \code{TRUE} the function creates and
interactive menu in case the best match is further than \code{max_string_dist}.
Check details and examples to understand why this parameter is important.
Ignored if \code{place} is not a character vector since the matching is
performed through a spatial operation.}

\item{verbose}{Boolean. If \code{TRUE} the function prints informative messages.}
}
\value{
A
}
\description{
Title
}
\details{
ABC
}
\examples{
1 + 1
}
