% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{osmext_get}
\alias{osmext_get}
\title{Title}
\usage{
osmext_get(
  place,
  provider = "geofabrik",
  match_by = "name",
  format = "pbf",
  max_string_dist = 1,
  interactive_ask = FALSE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{place}{String for the geographic zone that should be downloaded and
loaded.}

\item{provider}{Which provider should be used to download the data associated
to the mathced \code{place}? For the moment we support only
\href{http://download.geofabrik.de/}{\code{"geofabrik"}}.}

\item{match_by}{Which column of the provider data should be used for matching
the input place? The default value is "name". Check details and examples to
understand why this parameter is important.}

\item{format}{File format that will be downloaded from the chosen provider.
For the moment only the .osm.pbf format for download}

\item{max_string_dist}{What is the maximum distance in fuzzy matching to
tolerate before asking the user to select which zone to download?}

\item{interactive_ask}{What to do is the closest match is further than max_string_dist?}

\item{verbose}{Should the user be asked before downloading the file?}

\item{...}{Additional arguments passed to \code{\link[sf:st_read]{sf::st_read()}}}
}
\value{
A
}
\description{
Title
}
\details{
ABC
}
\examples{
1 + 1
}
