% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match.R
\name{osmext_match}
\alias{osmext_match}
\alias{osmext_match.default}
\alias{osmext_match.character}
\title{Match input place with a url}
\usage{
osmext_match(place, ...)

\method{osmext_match}{default}(place, ...)

\method{osmext_match}{character}(
  place,
  ...,
  provider = "geofabrik",
  match_by = "name",
  max_string_dist = 1,
  interactive_ask = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{place}{A}

\item{...}{B}

\item{provider}{C}

\item{match_by}{D}

\item{max_string_dist}{E}

\item{interactive_ask}{F}

\item{verbose}{G}
}
\value{
ABC
}
\description{
Match input place with a url
}
\examples{
1 + 1
}
