% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_pbf.R
\name{read_pbf}
\alias{read_pbf}
\title{Read pbf files with additional attributes}
\usage{
read_pbf(
  dsn,
  layer = "lines",
  key = NULL,
  value = NULL,
  selected_columns = "*",
  ini_file = NULL,
  attributes = make_additional_attributes(layer = layer),
  append = TRUE,
  vectortranslate_options = NULL
)
}
\arguments{
\item{dsn}{The location of the file}

\item{layer}{Which layer to read in? One of
"points"           "lines"            "multipolygons"    "multilinestrings" or "other_relations"}

\item{key}{Character string defining the key values to subset the data from, e.g. \code{"highway"}}

\item{value}{The value(s) the \code{key} can take, e.g. \code{"cycleway"}}

\item{selected_columns}{The columns to return in the output}

\item{ini_file}{A modified version of
\url{https://github.com/OSGeo/gdal/raw/master/gdal/data/osmconf.ini} If NULL (the
default), then it's created using \code{\link[=make_ini_attributes]{make_ini_attributes()}} function.}

\item{attributes}{Vector of character strings naming attributes to import}

\item{append}{Should the columns named in \code{attributes} be appended to the default columns?
\code{TRUE} by default.}

\item{vectortranslate_options}{Character vector that specify the options
passed to ogr2ogr. See details.}
}
\description{
Read pbf files (typically downloaded with \code{\link[=get_geofabrik]{get_geofabrik()}}) translating
the \verb{.osm.pbf`` file into a }.gpkg`` format. See details and the discussion
in \url{https://github.com/OSGeo/gdal/issues/2100}
}
\examples{
\donttest{
pbf_url = geofabrik_zones$pbf_url[geofabrik_zones$name == "Isle of Wight"]
f = file.path(tempdir(), "test.osm.pbf")
download.file(pbf_url, f)
# testing read_sf
sf::st_layers(f)
res = sf::read_sf(f) # works
res = sf::read_sf(f, query = "select * from lines") # works
res = sf::read_sf(f, query = "select * from multipolygons") # works
res = read_pbf(f, layer = "multipolygons")
q = "select * from lines where highway = 'cycleway'"
res_cycleways = sf::read_sf(f, layer = "lines", query = q)
res_cycleways = read_pbf(f, key = "highway", value = "cycleway") # more concise
res = read_pbf(f)
names(res)
res = read_pbf(f, layer = "points")
names(res)
res = read_pbf(f, selected_columns = "highway") # only return highway column
names(res)
res_cycleway = res = read_pbf(f, layer = "lines", key = "highway", value = "cycleway")
plot(res_cycleway)
# uncomment to get big dataset
# f_en = gf_filename("England")
# u_en = geofabrik_zones$pbf_url[geofabrik_zones$name == "England"]
# download.file(u_en, f_en)
# cycleway_en = read_pbf(f_en, layer = "lines", key = "highway", value = "cycleway")
# plot(cycleway_en$geometry)
# pryr::object_size(cycleway_en)
}
}
