% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_pbf.R
\name{read_pbf}
\alias{read_pbf}
\title{Read pbf files with additional attributes}
\usage{
read_pbf(
  dsn,
  layer = "lines",
  key = NULL,
  value = NULL,
  selected_columns = "*",
  attributes = make_additional_attributes(layer = layer),
  ini_file = NULL,
  append = TRUE
)
}
\arguments{
\item{dsn}{The location of the file}

\item{layer}{Which layer to read in? One of
"points"           "lines"            "multipolygons"    "multilinestrings" or "other_relations"}

\item{key}{Character string defining the key values to subset the data from, e.g. `"highway"`}

\item{value}{The value(s) the `key` can take, e.g. `"cycleway"`}

\item{selected_columns}{The columns to return in the output}

\item{attributes}{Vector of character strings naming attributes to import}

\item{ini_file}{A modified version of https://github.com/OSGeo/gdal/raw/master/gdal/data/osmconf.ini}

\item{append}{Should the columns named in `attributes` be appended to the default columns?
`TRUE` by default.}
}
\description{
Read pbf files with additional attributes
}
\examples{
\donttest{
pbf_url = geofabric_zones$pbf_url[geofabric_zones$name == "Isle of Wight"]
f = file.path(tempdir(), "test.osm.pbf")
download.file(pbf_url, f)
# testing read_sf
sf::st_layers(f)
res = sf::read_sf(f) # works
res = sf::read_sf(f, query = "select * from lines") # works
res = sf::read_sf(f, query = "select * from multipolygons") # works
res = read_pbf(f, layer = "multipolygons")
q = "select * from lines where highway = 'cycleway'"
res_cycleways = sf::read_sf(f, layer = "lines", query = q)
res_cycleways = read_pbf(f, key = "highway", value = "cycleway") # more concise
res = read_pbf(f)
names(res)
res = read_pbf(f, layer = "points")
names(res)
res = read_pbf(f, selected_columns = "highway") # only return highway column
names(res)
res_cycleway = res = read_pbf(f, layer = "lines", key = "highway", value = "cycleway")
plot(res_cycleway)
# uncomment to get big dataset
# f_en = gf_filename("England")
# u_en = geofabric_zones$pbf_url[geofabric_zones$name == "England"]
# download.file(u_en, f_en)
# cycleway_en = read_pbf(f_en, layer = "lines", key = "highway", value = "cycleway")
# plot(cycleway_en$geometry)
# pryr::object_size(cycleway_en)
}
}
