% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-network-graph.R
\name{tidy_oneway}
\alias{tidy_oneway}
\title{Tidy the oneway values in a osm network}
\usage{
tidy_oneway(net_raw, quiet = FALSE)
}
\arguments{
\item{net_raw}{a \code{sf} object representing a spatial network with the \code{oneway} and \code{junction} columns}

\item{quiet}{logical, whether to suppress messages. Default is FALSE.}
}
\value{
An \code{sf} object with standardised oneway values
}
\description{
This helper function standardises the oneway values in a osm network.
It also applies the implied \code{oneway} tag restriction based on the \code{junction}
tag values if specified.
}
\details{
For more information on the implied oneway restriction, see \href{https://wiki.openstreetmap.org/wiki/Key:oneway#Implied_oneway_restriction}{wiki.openstreetmap.org}.
}
\examples{
\dontrun{
sf_net <- osmextract::oe_get_network(place = "ITS Leeds", mode = "driving")
sf_net_tidy <- tidy_oneway(sf_net)
}
}
